fix: changed input type from float to int

+++ main.py
@@ -3,15 +3,15 @@ import Simplex
 
 
 def main():
-    C = np.array(list(map(int, input("Enter the coefficients of the function:\n").split())))
+    C = np.array(list(map(float, input("Enter the coefficients of the function:\n").split())))
     A = []
     num_of_constraints = int(input("How many constraints?\n"))
     print("Enter the coefficients of variables:")
     for i in range(num_of_constraints):
-        A.append(list(map(int, input().split())))
+        A.append(list(map(float, input().split())))
     A = np.array(A)
     print("Enter the free variables of the constraints:")
-    b = np.array(list(map(int, input().split())))
+    b = np.array(list(map(float, input().split())))
 
     approximation_accuracy = int(input("Enter the approximation accuracy:\n"))
     flag = int(input("maximize or minimize? [1/2]: "))
